Update SDK to 2024.1 (#105)
* Update SdkVersion in Directory.Build.props

The SdkVersion in Directory.Build.props has been updated from version 2023.3.2 to 2024.1.0-eap01. This update is in line with the latest SDK changes and aims to maintain compatibility.

* Update SDK version and target framework

The SDK version has been updated from the Early Access Preview version to the stable version. The .NET target framework version in the _build.csproj file has been updated to 8.0. Also, the versions of the Nuke.Common, dotnet-sonarscanner, NuGet.CommandLine, and NUnit.ConsoleRunner packages have been updated. Lastly, the '.gitignore' file was updated to ignore GitHub Copilot chat sessions.

* Update SdkVersion to 2024.1.4

The SdkVersion in the `Directory.Build.props` file has been updated from 2024.1.0 to 2024.1.4. This version bump ensures the project uses the latest features and bug fixes from the SDK.

* Update MicrosoftLoggingPackage version and add TestNet60 attribute

The version of the MicrosoftLoggingPackage constant in the NugetPackages.cs file was updated from 1.0.0 to 6.0.0. Additionally, the TestNet60 attribute was added to the ContextualLoggerConstructorAnalyzerTests class in the ContextualLoggerConstructorAnalyzerTests.cs file.