Upgrade SDK to 2024.2 (#110)
* Update code style and upgrade SDK version

This commit introduces an updated code style in TemplateParameterNameAttributeProvider.cs and _build.csproj.DotSettings. The style of the code has been simplified to a more straightforward and cleaner format.

* Update ShellComponent attribute in StructuredLoggingWikiDataProvider

The ShellComponent attribute in the StructuredLoggingWikiDataProvider.cs file has been updated. The instantiation was changed to DemandAnyThreadSafe to ensure the component's safety across various threads. This change is important to maintain thread safety and avoid potential race conditions or deadlocks.

* Update SdkVersion in Directory.Build.props

The SdkVersion property in Directory.Build.props file has been updated. The version has been changed from 2024.2.0-eap06 to 2024.2.0-eap07. This update is a part of the regular sdk upgrade process.

* Update SDK version in Directory.Build.props

Changed the SDK version from 2024.2.0-eap07 to 2024.2.0-eap09. This ensures the project uses the latest SDK features and bug fixes.

* Update SDK version to 2024.2.0-rc01

Updated the SDK version in Directory.Build.props from 2024.2.0-eap09 to 2024.2.0-rc01. This update ensures we are using the release candidate version for better stability and compatibility.

* Update SdkVersion to the stable release version

The SdkVersion has been changed from a release candidate (2024.2.0-rc01) to the stable version (2024.2.0) in the project properties. This ensures more reliability and stability for the builds.